chore: Moves all test utilities in a shared folder
