+++ test/integration/rules/extendedTests.ts
@@ -1,5 +1,5 @@
-import { assertLint } from '../assertLint';
-import { lintJavaScript } from '../esLintTester';
+import { assertLint } from '../../shared/assertLint';
+import { lintJavaScript } from '../../shared/esLintTester';
 
 suite('extended/', (): void => {
   test('consistent-err-names: disallows err as postfix.', async (): Promise<void> => {
