+++ test/shared/assertLint.ts
